// frontend-angular/src/proto/driver.ts
export interface DriverProfile {
  driverId: string;
  userId: string;
  vehicleNo: string;
  capacity: number;
  routes: RoutePlan[];
  model: string;
  color: string;
}

export interface RoutePlan {
  routeId: string;
  driverId: string;
  stops: RouteStop[];
  finalAreaId: string;
  createdAt: Date | undefined;
}

export interface RouteStop {
  sequence: number;
  areaId: string;
  isStation: boolean;
  arrivalOffsetMinutes: number;
}
// ...
export const DriverProfile: MessageFns<DriverProfile> = {
  encode(message: DriverProfile, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.driverId !== "") {
      writer.uint32(10).string(message.driverId);
    }
    if (message.userId !== "") {
      writer.uint32(18).string(message.userId);
    }
    if (message.vehicleNo !== "") {
      writer.uint32(26).string(message.vehicleNo);
    }
    if (message.capacity !== 0) {
      writer.uint32(32).int32(message.capacity);
    }
    for (const v of message.routes) {
      RoutePlan.encode(v!, writer.uint32(42).fork()).join();
    }
    if (message.model !== "") {
      writer.uint32(50).string(message.model);
    }
    if (message.color !== "") {
      writer.uint32(58).string(message.color);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): DriverProfile {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseDriverProfile();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.driverId = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.userId = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.vehicleNo = reader.string();
          continue;
        }
        case 4: {
          if (tag !== 32) {
            break;
          }

          message.capacity = reader.int32();
          continue;
        }
        case 5: {
          if (tag !== 42) {
            break;
          }

          message.routes.push(RoutePlan.decode(reader, reader.uint32()));
          continue;
        }
        case 6: {
          if (tag !== 50) {
            break;
          }

          message.model = reader.string();
          continue;
        }
        case 7: {
          if (tag !== 58) {
            break;
          }

          message.color = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): DriverProfile {
    return {
      driverId: isSet(object.driverId) ? globalThis.String(object.driverId) : "",
      userId: isSet(object.userId) ? globalThis.String(object.userId) : "",
      vehicleNo: isSet(object.vehicleNo) ? globalThis.String(object.vehicleNo) : "",
      capacity: isSet(object.capacity) ? globalThis.Number(object.capacity) : 0,
      routes: globalThis.Array.isArray(object?.routes) ? object.routes.map((e: any) => RoutePlan.fromJSON(e)) : [],
      model: isSet(object.model) ? globalThis.String(object.model) : "",
      color: isSet(object.color) ? globalThis.String(object.color) : "",
    };
  },

  toJSON(message: DriverProfile): unknown {
    const obj: any = {};
    if (message.driverId !== "") {
      obj.driverId = message.driverId;
    }
    if (message.userId !== "") {
      obj.userId = message.userId;
    }
    if (message.vehicleNo !== "") {
      obj.vehicleNo = message.vehicleNo;
    }
    if (message.capacity !== 0) {
      obj.capacity = Math.round(message.capacity);
    }
    if (message.routes?.length) {
      obj.routes = message.routes.map((e) => RoutePlan.toJSON(e));
    }
    if (message.model !== "") {
      obj.model = message.model;
    }
    if (message.color !== "") {
      obj.color = message.color;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<DriverProfile>, I>>(base?: I): DriverProfile {
    return DriverProfile.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<DriverProfile>, I>>(object: I): DriverProfile {
    const message = createBaseDriverProfile();
    message.driverId = object.driverId ?? "";
    message.userId = object.userId ?? "";
    message.vehicleNo = object.vehicleNo ?? "";
    message.capacity = object.capacity ?? 0;
    message.routes = object.routes?.map((e) => RoutePlan.fromPartial(e)) || [];
    message.model = object.model ?? "";
    message.color = object.color ?? "";
    return message;
  },
};

function createBaseRoutePlan(): RoutePlan {
  return { routeId: "", driverId: "", stops: [], finalAreaId: "", createdAt: undefined };
}

export const RoutePlan: MessageFns<RoutePlan> = {
  encode(message: RoutePlan, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.routeId !== "") {
      writer.uint32(10).string(message.routeId);
    }
    if (message.driverId !== "") {
      writer.uint32(18).string(message.driverId);
    }
    for (const v of message.stops) {
      RouteStop.encode(v!, writer.uint32(26).fork()).join();
    }
    if (message.finalAreaId !== "") {
      writer.uint32(34).string(message.finalAreaId);
    }
    if (message.createdAt !== undefined) {
      Timestamp.encode(toTimestamp(message.createdAt), writer.uint32(42).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): RoutePlan {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseRoutePlan();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.routeId = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.driverId = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.stops.push(RouteStop.decode(reader, reader.uint32()));
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.finalAreaId = reader.string();
          continue;
        }
        case 5: {
          if (tag !== 42) {
            break;
          }

          message.createdAt = fromTimestamp(Timestamp.decode(reader, reader.uint32()));
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): RoutePlan {
    return {
      routeId: isSet(object.routeId) ? globalThis.String(object.routeId) : "",
      driverId: isSet(object.driverId) ? globalThis.String(object.driverId) : "",
      stops: globalThis.Array.isArray(object?.stops) ? object.stops.map((e: any) => RouteStop.fromJSON(e)) : [],
      finalAreaId: isSet(object.finalAreaId) ? globalThis.String(object.finalAreaId) : "",
      createdAt: isSet(object.createdAt) ? fromJsonTimestamp(object.createdAt) : undefined,
    };
  },

  toJSON(message: RoutePlan): unknown {
    const obj: any = {};
    if (message.routeId !== "") {
      obj.routeId = message.routeId;
    }
    if (message.driverId !== "") {
      obj.driverId = message.driverId;
    }
    if (message.stops?.length) {
      obj.stops = message.stops.map((e) => RouteStop.toJSON(e));
    }
    if (message.finalAreaId !== "") {
      obj.finalAreaId = message.finalAreaId;
    }
    if (message.createdAt !== undefined) {
      obj.createdAt = message.createdAt.toISOString();
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<RoutePlan>, I>>(base?: I): RoutePlan {
    return RoutePlan.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<RoutePlan>, I>>(object: I): RoutePlan {
    const message = createBaseRoutePlan();
    message.routeId = object.routeId ?? "";
    message.driverId = object.driverId ?? "";
    message.stops = object.stops?.map((e) => RouteStop.fromPartial(e)) || [];
    message.finalAreaId = object.finalAreaId ?? "";
    message.createdAt = object.createdAt ?? undefined;
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };
// ...
export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}